Warehouses become ordered stacks, turning where you put things into the
decision the storage axis never had.

Storing costs per unit and holding costs per unit per cycle; reordering
costs per position moved and unstacking per item of depth. Retrieval is
destructive - reaching a buried cargo brings everything above it out with
it, and putting those back is a fresh paid decision, so the cost of digging
is not a formula but a consequence. Storing in delivery order nets about 5%
more than storing against it, carried mostly by re-storing what came up with
the target.

Storing is a choice, not a toll: extracted ore can ship and be delivered
without ever entering a warehouse.

This also kills the accounting that was broken before it. Three routes wrote
occupancy with no reference to any cargo - one accepted an arbitrary number
and would zero a full warehouse - and delivery never released space at all,
so with 1000 capacity against 1484 ore the world eventually bricked itself.
Occupancy is now a function of what is stacked and cannot be written on its
own.

The same defect recurred five times behind five different doors, always an
operation the world logged invalid after it had already mutated something:
truncating the stack before releasing space, debiting after stacking,
stacking the first of a duplicated pair, overflowing capacity mid-batch, and
requiring in-hand so hard that the warehouse became mandatory. None were
caught by a test - the scenarios fabricated cargo in its final state and
routed around the paths that break. Each is now pinned.

Measured limits are recorded in
docs/superpowers/revisoes/2026-08-10-revisao-final-armazenagem-posicional.md:
choosing the right ordering key is worth 0.02% while deciding to organise at
all is worth 2.2%, and custo_de_manutencao_por_unidade at 0.5 inverts the
ordering margin entirely.

241 tests.

+++ mundo/dominio/energia.py
@@ -15,28 +15,28 @@ class GerenciadorDeEnergia:
     def __init__(
         self,
         centrais: list[str],
-        energia_inicial_por_central: int = 10,
-        energia_total: int = 1000,
+        energia_inicial_por_central: float = 10,
+        energia_total: float = 1000,
     ) -> None:
         saldo_inicial_centrais = energia_inicial_por_central * len(centrais)
-        self._saldos: dict[str, int] = {central: energia_inicial_por_central for central in centrais}
+        self._saldos: dict[str, float] = {central: energia_inicial_por_central for central in centrais}
         self._saldos[self.RESERVA] = energia_total - saldo_inicial_centrais
 
-    def consultar_energia(self, central: str) -> int:
+    def consultar_energia(self, central: str) -> float:
         self._validar_central(central)
         return self._saldos[central]
 
-    def alocar_energia(self, origem: str, destino: str, quantidade: int) -> None:
+    def alocar_energia(self, origem: str, destino: str, quantidade: float) -> None:
         self._validar_quantidade(quantidade)
         if origem != self.RESERVA:
             raise PermissionError("Somente a Central de Missão pode alocar a partir da reserva")
         self._transferir(origem, destino, quantidade)
 
-    def redistribuir_energia(self, origem: str, destino: str, quantidade: int) -> None:
+    def redistribuir_energia(self, origem: str, destino: str, quantidade: float) -> None:
         self._validar_quantidade(quantidade)
         self._transferir(origem, destino, quantidade)
 
-    def revogar_energia(self, central: str, quantidade: int) -> None:
+    def revogar_energia(self, central: str, quantidade: float) -> None:
         self._validar_quantidade(quantidade)
         self._validar_central(central)
         if self._saldos[central] < quantidade:
@@ -44,14 +44,14 @@ class GerenciadorDeEnergia:
         self._saldos[central] -= quantidade
         self._saldos[self.RESERVA] += quantidade
 
-    def debitar(self, central: str, quantidade: int) -> None:
+    def debitar(self, central: str, quantidade: float) -> None:
         self._validar_quantidade(quantidade)
         self._validar_central(central)
         if self._saldos[central] < quantidade:
             raise EnergiaInsuficienteError(central)
         self._saldos[central] -= quantidade
 
-    def _transferir(self, origem: str, destino: str, quantidade: int) -> None:
+    def _transferir(self, origem: str, destino: str, quantidade: float) -> None:
         self._validar_central(origem)
         self._validar_central(destino)
         if self._saldos[origem] < quantidade:
@@ -63,6 +63,6 @@ class GerenciadorDeEnergia:
         if central not in self._saldos:
             raise CentralDesconhecidaError(central)
 
-    def _validar_quantidade(self, quantidade: int) -> None:
+    def _validar_quantidade(self, quantidade: float) -> None:
         if quantidade <= 0:
             raise ValueError("Quantidade deve ser positiva")
